requirements.txt
asgiref==3.2.5
cycler==0.10.0
Django==3.0.4
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
pandas==1.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
seaborn==0.9.0
django-heroku==0.3.1
gunicorn==19.9.0
